+++ app/src/main/res/layout/activity_diagnostic_list.xml
@@ -1,12 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.pedigreetechnologies.diagnosticview.DiagnosticList"
-    tools:showIn="@layout/app_bar_main">
+    tools:context="com.pedigreetechnologies.diagnosticview.DiagnosticList">
 
     <TextView
         android:id="@+id/d_select_text"
@@ -49,4 +46,4 @@
         android:layout_below="@android:id/list"
         android:textSize="11dp"
         />
-</android.support.constraint.ConstraintLayout>
+</RelativeLayout>
