+++ app/src/main/res/layout/app_bar_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.pedigreetechnologies.diagnosticview.DiagnosticList">
+    tools:context="com.pedigreetechnologies.diagnosticview.ViewPagerFragmentActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
@@ -18,8 +18,9 @@
             android:background="?attr/colorPrimary"
             app:popupTheme="@style/AppTheme.PopupOverlay" />
 
+
     </android.support.design.widget.AppBarLayout>
 
-    <include layout="@layout/activity_diagnostic_list" />
+    <include layout="@layout/viewpager_layout" />
 
 </android.support.design.widget.CoordinatorLayout>
