+++ app/src/main/res/layout/viewpager_layout.xml
@@ -1,5 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<android.support.constraint.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    app:layout_behavior="@string/appbar_scrolling_view_behavior"
+    tools:context="com.pedigreetechnologies.diagnosticview.ViewPagerFragmentActivity"
+    tools:showIn="@layout/app_bar_main">
+
+    <LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
@@ -11,3 +20,4 @@
         android:layout_height="fill_parent"
         />
 </LinearLayout>
+</android.support.constraint.ConstraintLayout>
